wingの追加、高さで測るように、警告の削除

+++ Game/Scene/ResultScene.h
@@ -81,7 +81,7 @@ private:
 	bool isStartScene_ = true;
 	static bool isViewingRanking_;
 
-	int score_ = FlyingTimer::flyingScor;
+	int score_ = (int)FlyingTimer::flyingScor;
 	SCORE_RANK rank_;
 
 	/*---------- debug ----------*/
